+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 # Current version
 
+# Version 18.01.24
+
+* Fix changing email
+* added news author in emails
+* Fixed events' start time in iCalc
+* Fixed remove student from queue when enrolling to lecture
+* Fixed issue with contextprocessors
+* Fixed editing course in offer
+
 # Version 18.01.05
 
 * Upgrade to Django 1.10